move showcase ui to tools/vyy-showcase
