+++ vuu-ui/showcase/package.json
@@ -21,6 +21,7 @@
     "@finos/vuu-data-ag-grid": "0.0.26",
     "@finos/vuu-filters": "0.0.26",
     "@finos/vuu-layout": "0.0.26",
+    "@finos/vuu-showcase": "0.0.1",
     "@finos/vuu-theme": "0.0.26",
     "@finos/vuu-utils": "0.0.26",
     "@salt-ds/core": "1.13.2",
