+++ vuu-ui/tools/vuu-showcase/src/Showcase.tsx
@@ -1,5 +1,7 @@
-import "@finos/vuu-icons/index.css";
-import "@finos/vuu-theme/index.css";
+// The Showcase chrome is always rendered with Vuu styling, hence we always need the Vuu theme
+import "./themes/vuu";
+
+import "./Showcase.css";
 
 import { AppRoutes } from "./AppRoutes";
 import { ExamplesModule } from "./showcase-utils";
